+++ qupath-core/src/main/java/qupath/lib/color/ColorDeconvolutionStains.java
@@ -27,11 +27,15 @@ import java.io.Externalizable;
 import java.io.IOException;
 import java.io.ObjectInput;
 import java.io.ObjectOutput;
+import java.text.NumberFormat;
+import java.util.Locale;
 import java.util.Map;
+import java.util.StringTokenizer;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import qupath.lib.color.StainVector.DEFAULT_STAINS;
 import qupath.lib.common.GeneralTools;
 
 
@@ -365,11 +369,12 @@ public class ColorDeconvolutionStains implements Externalizable {
 			if (i == 3 && stain.isResidual())
 				continue;
 			sb.append("\"Stain ").append(i).append("\" : \"").append(stain.getName()).append("\", ");
-			sb.append("\"Values ").append(i).append("\" : \"").append(stain.arrayAsString(nDecimalPlaces)).append("\", ");
+			sb.append("\"Values ").append(i).append("\" : \"").append(stain.arrayAsString(Locale.US, nDecimalPlaces)).append("\", ");
 		}
 		sb.append("\"Background\" : \"");
 		
-		sb.append(String.format( "%.Nf %.Nf %.Nf".replace("N", Integer.toString(nDecimalPlaces)), stains.getMaxRed(), stains.getMaxGreen(), stains.getMaxBlue()));
+		sb.append(" ").append(GeneralTools.arrayToString(Locale.US, new double[]{stains.getMaxRed(), stains.getMaxGreen(), stains.getMaxBlue()}, nDecimalPlaces));
+//		sb.append(String.format( "%.Nf %.Nf %.Nf".replace("N", Integer.toString(nDecimalPlaces)), stains.getMaxRed(), stains.getMaxGreen(), stains.getMaxBlue()));
 //		sb.append(stains.getMaxRed()).append(" ");
 //		sb.append(stains.getMaxGreen()).append(" ");
 //		sb.append(stains.getMaxBlue());
@@ -385,16 +390,85 @@ public class ColorDeconvolutionStains implements Externalizable {
 		if (map.isEmpty())
 			return null;
 		
-		StainVector stain1 = StainVector.parseStainVector(map.get("Stain 1"), map.get("Values 1"));
-		StainVector stain2 = StainVector.parseStainVector(map.get("Stain 2"), map.get("Values 2"));
+		StainVector stain1 = parseStainVector(Locale.US, map.get("Stain 1"), map.get("Values 1"));
+		StainVector stain2 = parseStainVector(Locale.US, map.get("Stain 2"), map.get("Values 2"));
 		StainVector stain3 = null;
 		if (map.containsKey("Stain 3"))
-			stain3 = StainVector.parseStainVector(map.get("Stain 3"), map.get("Values 3"));
+			stain3 = parseStainVector(Locale.US, map.get("Stain 3"), map.get("Values 3"));
 		
-		double[] background = StainVector.parseValues(map.get("Background"));
+		double[] background = parseStainValues(Locale.US, map.get("Background"));
 		
 		return new ColorDeconvolutionStains(map.get("Name"), stain1, stain2, stain3, background[0], background[1], background[2]);
 	}
+	
+	
+	public static StainVector parseStainVector(Locale locale, String name, String s) {
+		if (s == null)
+			return null;
+		s = s.trim();
+		if (s.length() == 0)
+			return null;
+		if (s.length() == 1) {
+			s = s.toLowerCase();
+			if (s.equals("h"))
+				return StainVector.makeDefaultStainVector(DEFAULT_STAINS.HEMATOXYLIN);
+			else if (s.equals("e"))
+				return StainVector.makeDefaultStainVector(DEFAULT_STAINS.EOSIN);
+			else if (s.equals("d"))
+				return StainVector.makeDefaultStainVector(DEFAULT_STAINS.DAB);
+		}
+		double[] vector = parseStainValues(locale, s);
+		if (vector != null)
+			return new StainVector(name, vector[0], vector[1], vector[2]);
+		// If we got here, we didn't find 3 numbers
+		return null;
+	}
+	
+	/**
+	 * Parses 3 values, suitable for use as a stain vector
+	 * @param s
+	 * @return
+	 */
+	public static double[] parseStainValues(Locale locale, String s) {
+		// Try to parse the stain vector
+		double[] vector = new double[3];
+		StringTokenizer tokenizer = new StringTokenizer(s, " \t\n\r\f[]");
+		int i = 0;
+		while (tokenizer.hasMoreTokens()) {
+			String token = tokenizer.nextToken();
+			if (token.endsWith(","))
+				token = token.substring(0, token.length()-1);
+			try {
+				vector[i] = NumberFormat.getInstance(locale).parse(token).doubleValue();// Double.parseDouble(token);
+			} catch (Exception e) {
+//				// Try this as a last resort...?
+				vector[i] = Double.parseDouble(token);				
+//				if (token.contains(",") && !token.contains(".")) {
+//					String newToken = token.replace(",", ".");
+//					vector[i] = Double.parseDouble(newToken);				
+//					logger.warn("Using emergency stain vector parsing! {} was replaced with {}", token, newToken);
+//				}
+			}
+			i++;
+			// If we got here, we found 3 numbers
+			if (i == 3)
+				return vector;
+		}
+		
+//		Matcher matcher = StainVector.pattern.matcher(s);
+//		double[] vector = new double[3];
+//		int i = 0;
+//		while (matcher.find()) {
+//			vector[i] = Double.parseDouble(matcher.group());
+//			i++;
+//			// If we got here, we found 3 numbers
+//			if (i == 3)
+//				return vector;
+//		}
+		// If we got here, we didn't find 3 numbers
+		return null;
+	}
+