+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## Version 0.1.0
+
+* Fixed bug in 'Show setup options' that made it difficult to return changed region settings back to the default (but any other change was ok)
+* Improved consistency of formatting used to display numbers based on other Locales
+* Switched default to US Locale
+* Removed pre-release notification
+* Switched build to request a system rather than user installation (mostly so as to automatically request admin privileges on Windows)
+
+
 ## Version 0.0.7
 
 * New 'Show setup options' dialog to encourage choosing important settings when QuPath is first used.